ensure proper AS7 logs are propagated plus disabling http spider for AS7 for now

+++ scripts/jbossws-qa.sh
@@ -36,8 +36,13 @@ startJBoss() {
 }
 
 copyJBossLogs() {
-  cp $JBOSS_HOME/server/$JBOSS_CONFIG/log/boot.log $WORKSPACE/jboss-boot.log
-  cp $JBOSS_HOME/server/$JBOSS_CONFIG/log/server.log $WORKSPACE/jboss-server.log
+  if [ -f $JBOSS_HOME/bin/run.sh ]; then
+    cp $JBOSS_HOME/server/$JBOSS_CONFIG/log/boot.log $WORKSPACE/jboss-boot.log
+    cp $JBOSS_HOME/server/$JBOSS_CONFIG/log/server.log $WORKSPACE/jboss-server.log
+  else
+    cp $JBOSS_HOME/standalone/log/boot.log $WORKSPACE/jboss-boot.log
+    cp $JBOSS_HOME/standalone/log/server.log $WORKSPACE/jboss-server.log
+  fi
 }
 
 copyTestLogs() {
@@ -46,17 +51,27 @@ copyTestLogs() {
 }
 
 removeJBossLogs() {
-  rm -f $JBOSS_HOME/server/$JBOSS_CONFIG/log/boot.log
-  rm -f $JBOSS_HOME/server/$JBOSS_CONFIG/log/server.log
+  if [ -f $JBOSS_HOME/bin/run.sh ]; then
+    rm -f $JBOSS_HOME/server/$JBOSS_CONFIG/log/boot.log
+    rm -f $JBOSS_HOME/server/$JBOSS_CONFIG/log/server.log
+  else
+    rm -f $JBOSS_HOME/standalone/log/boot.log
+    rm -f $JBOSS_HOME/standalone/log/server.log
+  fi
 }
 
 ensureRunningJBoss() {
-  $SCRIPTS_DIR/http-spider.sh $JBOSS_BIND_ADDRESS:8080 $WORKSPACE
-  if [ -e $WORKSPACE/spider.failed ]; then
-    tail -n 100 $JBOSS_HOME/server/$JBOSS_CONFIG/log/server.log
-    stopJBoss
-    copyJBossLogs
-    exit 1
+  if [ -f $JBOSS_HOME/bin/run.sh ]; then
+    $SCRIPTS_DIR/http-spider.sh $JBOSS_BIND_ADDRESS:8080 $WORKSPACE
+    if [ -e $WORKSPACE/spider.failed ]; then
+      tail -n 100 $JBOSS_HOME/server/$JBOSS_CONFIG/log/server.log
+      stopJBoss
+      copyJBossLogs
+      exit 1
+    fi
+  else
+    echo "There's no admin console for AS7 yet"
+    sleep 20
   fi
 }
